% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue.R
\name{gross_maple_revenue}
\alias{gross_maple_revenue}
\title{Gross maple revenue}
\usage{
gross_maple_revenue(syrup_volume)
}
\arguments{
\item{syrup_volume}{vector of syrup volumes (gallons) produced per tree}
}
\value{
vector of gross revenues (USD present value) attributable to each
  tree
}
\description{
Gross maple revenue
}
