% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{sap_collection_costs}
\alias{sap_collection_costs}
\title{Sap collection costs}
\usage{
sap_collection_costs(
  taps_per_acre,
  wage = 25,
  life_years = 10,
  real_interest = 0.05
)
}
\arguments{
\item{taps_per_acre}{number of maple taps per acre}

\item{wage}{wage for woods work (USD/hr)}

\item{life_years}{number of years to amortize tubing, etc.}

\item{real_interest}{interest rate after inflation (for capital recovery)}
}
\value{
annual per acre cost (USD) of sap collection, exclusive of land
  rent, syrup production, marketing, etc.
}
\description{
Sap collection costs
}
\references{
After Huyler 2000 (USFS Res Pap NE-712); UVM 'Business of Sap'
  benchmarking slides (2020); Childs 2019 'Estimating costs for installing a
  maple tubing system'; and Cornell "Beginer's Notebook" & tap density
  guidance (2019)
}
