% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{sugar_content}
\alias{sugar_content}
\title{Maple sugar content}
\usage{
sugar_content(dbh, spp)
}
\arguments{
\item{dbh}{numeric vector of trees' diameters at breast height (inches)}

\item{spp}{character vector of trees' species}
}
\value{
vector of sap sugar contents (brix) for each tree
}
\description{
sap sugar content for individual trees
}
\references{
after Rademacher et al 2023; Larochelle et al 1998; Wild and Yana 2015
}
