% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{syrup_yield}
\alias{syrup_yield}
\title{Maple syrup yield}
\usage{
syrup_yield(dbh, spp)
}
\arguments{
\item{dbh}{numeric vector of trees' diameters at breast height (inches)}

\item{spp}{character vector of trees' species}
}
\value{
vector of syrup volumes (gallons) for each tree
}
\description{
gallons syrup produced per season for individual trees
}
\references{
after Rademacher et al 2023, but simplified as a linear relationship
}
